+++ harjus-port/plannedexecution.h
@@ -6,14 +6,14 @@
 class PlannedExecution
 {
 public:
-  double totalProfit() const { return mTotalProfit; }
+  long double totalProfit() const { return mTotalProfit; }
   void update();
   QJsonObject toJson() const;
-  PlannedExecution(std::vector<TradingSymbol> trades, double relativePrice, double commission) : mTrades{trades}, mTotalProfit{0}, mRelativePrice(relativePrice), mCommission(commission) {};
+  PlannedExecution(std::vector<TradingSymbol> trades, long double relativePrice, long double commission) : mTrades{trades}, mTotalProfit{0}, mRelativePrice(relativePrice), mCommission(commission) {};
 
 private:
   std::vector<TradingSymbol> mTrades;
-  double mTotalProfit;
-  double mRelativePrice;
-  double mCommission;
+  long double mTotalProfit;
+  long double mRelativePrice;
+  long double mCommission;
 };
